backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0
Werkzeug==3.0.3
